updated a* code and logic

+++ Final Project/Assets/Scripts/DataModel/Soldier.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 public abstract class Soldier : Character
 {
     public bool IsAlive { get; private set; }
-    public int Helath { get; private set; }
+    public float Helath { get; private set; }
     public List<Weapon> Weapons { get; set; }
